Track5  java8  cleanup IDE DesktopSupport/TaskbarSupport (#231)
* Update IDEDesktopSupport.java

* TaskbarSupport/DesktopSupport remove Java 8 code

- Delete legacy org.sikuli.support.ide.IDEDesktopSupport (moved/reworked), update SikulixIDE imports and initialization to use the revised idesupport package (Java 8 code removed)
- activated to show about
- Replace reflective Java8 taskbar handling with java.awt.Taskbar API, rename method to setTaskbarIcon and simplify icon setting logic.
- Add new gecko_cyclope.png icon resource used for the taskbar.
